requirements.txt
streamlit
pandas
numpy
opencv-python
Pillow
scikit-learn
tensorflow
